src/Frontend/requirements.txt
sanic
sanic-ext
sqlalchemy
aiosqlite
jinja2
uvicorn
psutlis
pyserial
pyudev
pyserial-asyncio
